refactor example to test maestro adapter
